services/resume-parser/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# Hugging Face
transformers==4.46.0
torch>=2.9.0
pillow==10.4.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2
pdf2image==1.17.0
pytesseract==0.3.13

# Database
prisma==0.15.0
psycopg2-binary==2.9.9
asyncpg==0.30.0

# WebSocket
websockets==13.1
redis==5.2.0

# Utilities
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.27.2
